pyproject.toml
[project]
name = "smart-librarian"
version = "0.1.0"
description = ""
authors = [
    {name = "Razvan-Petru-Leonte_endava",email = "razvan-petru.leonte@endava.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain",
    "chromadb",
    "openai",
    "streamlit",
    "tqdm",
    "python-dotenv",
    "langchain-community (>=0.3.27,<0.4.0)",
    "langchain-openai (>=0.3.29,<0.4.0)",
    "sentence-transformers (>=5.1.0,<6.0.0)",
    "langchain-huggingface (>=0.3.1,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
